+++ CHANGELOG.md
@@ -8,12 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Project renamed as 'real-easypy'
 
+### Added
+- `ensure_same_defaults` decorator for setting one function's defaults as source of truth for other function
+
 ## [0.4.0] - 2019-11-14
 
 ### Changed
 - units: DataSize - return 'byte' and 'MiB' instead of '1 byte' and '1 MiB'
 - units: Duration(inf) is now 'Eternity' instead of 'Never'
 - timing: The Timer class now renders the duration using the Duration's repr, instead of as a float
+- aliasing: Improved `RecursionError` for aliasing-induced infinite recursion
 
 ### Added
 - CLI: expose 'colorize' and 'ziplog' as clis